Move exceptions.py to utils/
